As my current interest is to read PR226 data, I am streamlining the parameters towards analyzing that data. Also, the x1offset list in the BOR routine is typical of PR226 runs
------------------------------------------------------------------

I added the option for setupchannel2wierXUXold in the main.c initilization routine.

------------------------------------------------------------------
Changes to ODB:
Took out the xcorr parameters, as these are done in the config file
At the same time I deleted the odb-connected pieces of code from
CalcCorrX in Focalplane.c

No longer is the inter-VDC separation numbers z_x1x2 and x_x1x2 read from the
ODB (in the Analayzer/Parameters/Global directory).
These are not found inside the config file in the VDC section

------------------------------------------------------------------
The routines to calculate the FP angles (vertical and horizontal)
from X1 and X2 was moved to FocalPlane.c from main.c,
and it was changed from type void to double.

Then the old CalcThSCAT routine was removed, and the new CalcThetaPrime
and CalcPhiPrime was renames to CalcThetaScat and CalcPhiScat
and some variables was renamed in main.c to have structure thetaSCAT instead of ThSCAT
etc.