Added t_X1posO
Instead of adding the offset to t_X1pos, I created a new variable called t_X1posO. This prevents confusion
on offline analysis, since the X1pos used in the angle calculations DOES NOT have the offset built into it
(and it SHOULD NOT have it either!) but the ttree variable X1pos does. If the offset is big this causes unnecessary confusion
since e.g. the following commands will not show the same results in the previous form of the code:
DATA->Draw("thetaSCAT:X1pos>>hThSCATvsX1",CUTbasic,"")
DATA->Draw("thetaFP*(-0.6495 -0.00003157*X1pos) + (23.63-0.003341*X1pos):X1pos>>hThSCATvsX1",CUTbasic,"")
(assuming these thefp-to-thscat parameters are in the config file)
In the first command the thetaSCAT was calculated in the code without any offset correction for X1pos.
In the second command we involve X1pos that already has the offset correction built into it.

In the changes submitted here both yields the same results.